Refactored tasks files from project & core
